Add GitHubBase module
